requirements.txt
asgiref==2.2.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
certifi==2018.1.18
channels==2.0.2
chardet==3.0.4
constantly==15.1.0
daphne==2.1.0
defusedxml==0.5.0
Django==2.0.3
django-allauth==0.35.0
django-bootstrap-form==3.3
django-debug-toolbar==1.9.1
django-taggit==0.22.2
django-uuidfield==0.5.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
oauthlib==2.0.6
psycopg2==2.7.4
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
Twisted==17.9.0
txaio==2.9.0
urllib3==1.22
zope.interface==4.4.3
